requirements.txt
# main dependencies
Django<1.9
six
# for product advertising api
beautifulsoup4==4.3.2
bottlenose>=0.6.2
celery>=3
python-dateutil
# for pm api
djangorestframework>=3.0.4
# for graphs
pygal>=1.5.1
lxml
# pygal png output
CairoSVG
tinycss
cssselect
